+++ NHotPhrase.sln
@@ -33,6 +33,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHotPhrase.WindowsForms.Tes
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHotPhrase.Wpf.Demo", "samples\NHotPhrase.Wpf.Demo\NHotPhrase.Wpf.Demo.csproj", "{B2464211-38B4-4CCB-9F88-1E6AD03A2A3C}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHotPhrase.Wpf", "src\NHotPhrase.Wpf\NHotPhrase.Wpf.csproj", "{B0256D9C-31AA-489E-A597-1BB6FE609F1B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -89,6 +91,14 @@ Global
 		{B2464211-38B4-4CCB-9F88-1E6AD03A2A3C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B2464211-38B4-4CCB-9F88-1E6AD03A2A3C}.Release|x64.ActiveCfg = Release|x64
 		{B2464211-38B4-4CCB-9F88-1E6AD03A2A3C}.Release|x64.Build.0 = Release|x64
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Debug|x64.Build.0 = Debug|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Release|x64.ActiveCfg = Release|Any CPU
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -100,6 +110,7 @@ Global
 		{9169DAE9-D879-4F88-9D4C-622B9204745F} = {90C6005F-8801-41EF-AE11-9B0C6483025E}
 		{AD8671BA-BE83-44D0-8500-F32DE7E6181A} = {FBA6085F-DD7E-4E18-9BC8-8F709157EC33}
 		{B2464211-38B4-4CCB-9F88-1E6AD03A2A3C} = {CEE3E414-E343-43C9-8CB9-72CA02FD0299}
+		{B0256D9C-31AA-489E-A597-1BB6FE609F1B} = {FBA6085F-DD7E-4E18-9BC8-8F709157EC33}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CACB69F6-E9A2-4B9C-84DC-6A91978A571B}
