+++ samples/NHotPhrase.Wpf.Demo/MainWindow.xaml
@@ -1,13 +1,7 @@
 ﻿<Window x:Class="NHotPhrase.Wpf.Demo.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="MainWindow" Height="300" Width="300">
-    <Window.InputBindings>
-        <KeyBinding Gesture="Ctrl+Alt+N" Command="{Binding NegateCommand}"
-                    HotPhraseManager.RegisterGlobalHotkey="True" />
-        <KeyBinding Gesture="Ctrl+Alt+T" Command="{Binding TestCommand}"
-                    HotPhraseManager.RegisterGlobalHotkey="True" />
-    </Window.InputBindings>
+        Title="MainWindow" Height="600" Width="350">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
@@ -16,16 +10,22 @@
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
         <CheckBox Grid.Row="0" Content="Global hotkeys enabled" IsChecked="{Binding IsHotkeyManagerEnabled}" />
         <TextBlock Grid.Row="1" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="48" Text="{Binding Value}" />
+        
+        <TextBlock Grid.Row="2" HorizontalAlignment="Left" Text="Toggle Global HotPhrases" />
         <TextBlock Grid.Row="2" HorizontalAlignment="Left" Text="Increment" />
         <TextBlock Grid.Row="3" HorizontalAlignment="Left" Text="Decrement" />
-        <TextBlock Grid.Row="4" HorizontalAlignment="Left" Text="Negate" />
-        <TextBlock Grid.Row="5" HorizontalAlignment="Left" Text="Test" />
-        <TextBlock Grid.Row="2" HorizontalAlignment="Right" Text="{Binding IncrementHotkey}" />
-        <TextBlock Grid.Row="3" HorizontalAlignment="Right" Text="{Binding DecrementHotkey}" />
-        <TextBlock Grid.Row="4" HorizontalAlignment="Right" Text="Ctrl+Alt+N" />
-        <TextBlock Grid.Row="5" HorizontalAlignment="Right" Text="Ctrl+Alt+T" />
+        <TextBlock Grid.Row="4" HorizontalAlignment="Left" Text="Send text" />
+        <TextBlock Grid.Row="5" HorizontalAlignment="Left" Text="Single wildcard digit after hot phrase" />
+        <TextBlock Grid.Row="6" HorizontalAlignment="Left" Text="Double wildcard alphanumeric hot phrase" />
+
+        <TextBlock Grid.Row="2" HorizontalAlignment="Right" Text="Right Ctrl, Right Ctrl, Right Ctrl" />
+        <TextBlock Grid.Row="3" HorizontalAlignment="Right" Text="Ctrl, Shift, Alt" />
+        <TextBlock Grid.Row="4" HorizontalAlignment="Right" Text="Caps Lock, Caps Lock, W, R, G" />
+        <TextBlock Grid.Row="5" HorizontalAlignment="Right" Text="Caps Lock, Caps Lock, N, #" />
+        <TextBlock Grid.Row="6" HorizontalAlignment="Right" Text="Caps Lock, Caps Lock, M, X, X" />
     </Grid>
 </Window>
