+++ src/NHotPhrase/Keyboard/ISendKeys.cs
@@ -5,8 +5,9 @@ namespace NHotPhrase.Keyboard
 {
     public interface ISendKeys
     {
+        int MillisecondsBetweenKeyPress { get; set; }
         bool SendKeysAndWait(PhraseActionRunState phraseActionRunState, List<PKey> keys);
-        bool SendKeysAndWait(string stringToSend, int millisecondThreadSleep = 2);
-        bool SendKeysAndWait(List<string> stringsToSend, int millisecondThreadSleep = 2);
+        bool SendKeysAndWait(string stringToSend, int millisecondThreadSleep = -1);
+        bool SendKeysAndWait(List<string> stringsToSend, int millisecondThreadSleep = -1);
     }
 }
